Add support for libkafs and libkopenafs
Probe for the libkafs and libkopenafs libraries and, if found, use
them in preference to our portability layer.  Still build the local
replacement for k_haspag if the libraries don't have it.

Remove the glue directory on each run of Build.PL so that we don't
have left-over glue from previous runs with different configurations.

Don't remove the entire glue directory on clean, just the object
files.  This means that we leave files behind after a build, but
there's no good way around that in Module::Build right now since it
doesn't support adding things to realclean.

Configuring this package also requires Perl 5.010.